backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Tuammar"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "paho-mqtt (>=2.1.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
